pyproject.toml
[project]
name = "llamachat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.3"
dependencies = [
    "ollama==0.4.8",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.16.0",
    "ruff>=0.11.12",
    "types-pyyaml>=6.0.12.20250516",
]

[tool.mypy]
ignore_missing_imports = true
